refactor: move kargs_hardening_common to align with current conventions
