+++ files/system/nvidia/usr/libexec/secureblue/remove_kargs_nvidia.py
@@ -10,12 +10,12 @@ import sys
 from typing import TYPE_CHECKING, Final
 
 if TYPE_CHECKING:
-    from files.system.usr.libexec.secureblue import kargs_hardening_common
+    from files.system.usr.libexec.secureblue.shared import kargs_hardening
 else:
-    import kargs_hardening_common
+    from shared import kargs_hardening
 
-IMAGE_NVIDIA_KARGS: Final = kargs_hardening_common.IMAGE_NVIDIA_KARGS
-apply_kargs: Final = kargs_hardening_common.apply_kargs
+IMAGE_NVIDIA_KARGS: Final = kargs_hardening.IMAGE_NVIDIA_KARGS
+apply_kargs: Final = kargs_hardening.apply_kargs
 
 
 def main() -> int:
